Moved helpful function `lineAndColumnOf` into utils.ts file
